+++ src/main/java/com/choongang/auction/streamingauction/service/ProductService.java
@@ -59,9 +59,9 @@ public class ProductService {
                 .category(category)
                 .categoryName(category.getCategoryType().name()) // 카테고리명 직접 설정
                 .member(member)
-                .startingPrice(BigDecimal.valueOf(dto.productStartPrice()))
-                .bidIncrease(BigDecimal.valueOf(dto.productBidIncrement()))
-                .buyNowPrice(BigDecimal.valueOf(dto.productBuyNowPrice()))
+                .startingPrice(Long.valueOf(dto.productStartPrice()))
+                .bidIncrease(Long.valueOf(dto.productBidIncrement()))
+                .buyNowPrice(Long.valueOf(dto.productBuyNowPrice()))
                 .images(new ArrayList<>())
                 .build();
